+++ Control/Joint/Schemes/TUT.hs
@@ -4,8 +4,8 @@ import Control.Joint.Core (type (:.), type (:=))
 import Control.Joint.Abilities.Interpreted (Interpreted (Primary, run))
 
 -- TODO: think about decomposing it on UT and TU
-newtype TUT t u t' a = TUT (t :. u :. t' := a)
+newtype TUT t t' u a = TUT (t :. u :. t' := a)
 
-instance Interpreted (TUT t u t') where
-	type Primary (TUT t u t') a = t :. u :. t' := a
+instance Interpreted (TUT t t' u) where
+	type Primary (TUT t t' u) a = t :. u :. t' := a
 	run (TUT x) = x